requirements.txt
ai4bharat-transliteration==1.1.3
alabaster==0.7.13
antlr4-python3-runtime==4.8
Babel==2.12.1
bitarray==2.7.3
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Cython==0.29.34
docutils==0.18.1
fairseq==0.12.2
filelock==3.12.0
Flask==2.3.2
Flask-Cors==3.0.10
gevent==22.10.2
greenlet==2.0.2
hydra-core==1.0.7
idna==3.4
imagesize==1.4.1
importlib-metadata==6.6.0
indic-nlp-library==0.91
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
lxml==4.9.2
MarkupSafe==2.1.2
mock==5.0.2
Morfessor==2.0.6
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
omegaconf==2.0.6
packaging==23.1
pandas==2.0.1
portalocker==2.7.0
progressbar2==4.2.0
protobuf==3.20.3
pyarrow==12.0.0
pycparser==2.21
pydload==1.0.9
Pygments==2.15.1
python-dateutil==2.8.2
python-utils==3.5.2
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
sacrebleu==2.3.1
sacremoses==0.0.53
six==1.16.0
snowballstemmer==2.2.0
Sphinx==6.2.1
sphinx-argparse==0.4.0
sphinx-rtd-theme==1.2.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sympy==1.12
tabulate==0.9.0
tensorboardX==2.6
torch==2.0.1
torchaudio==2.0.2
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
ujson==5.7.0
urduhack==0.1.4
urllib3==2.0.2
Werkzeug==2.3.4
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
